Fix warnings

+++ include/RED4ext/SharedPtr.hpp
@@ -46,7 +46,7 @@ struct RefCnt
     // Returns true if the strong refs count reached zero.
     bool DecRef()
     {
-        return InterlockedExchangeAdd(&strongRefs, -1) == 1;
+        return InterlockedExchangeAdd(&strongRefs, static_cast<uint32_t>(-1)) == 1;
     }
 
     void IncWeakRef()
